Moved Channel to database package
