+++ app/src/main/java/com/github/sdp/ratemyepfl/database/Channel.kt
@@ -1,7 +1,6 @@
-package com.github.sdp.ratemyepfl.model.messaging
+package com.github.sdp.ratemyepfl.database
 
-import com.github.sdp.ratemyepfl.database.ChannelListener
-import com.github.sdp.ratemyepfl.database.IChannel
+import com.github.sdp.ratemyepfl.model.messaging.Message
 import com.google.firebase.firestore.*
 import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
